+++ packages/core/src/SagaRegistry/effects.d.ts
@@ -14,7 +14,11 @@ export declare function take(
     sagaItem: SagaItem,
     pattern: Pattern<Action>
 ): TakeEffect;
-export declare function put(sagaItem: SagaItem, action: Action): PutEffect<any>;
+export declare function put(
+    sagaItem: SagaItem,
+    action: Action,
+    relativeDispatchPath: string
+): PutEffect<any>;
 export declare function select(
     sagaItem: SagaItem,
     selector?: (state: any, ...args: any[]) => any,
@@ -22,22 +26,26 @@ export declare function select(
 ): SelectEffect;
 export declare function takeEvery(
     sagaItem: SagaItem,
+    pattern: Pattern<Action>,
     saga: GeneratorFunction,
     ...args: any[]
 ): ForkEffect;
 export declare function takeLatest(
     sagaItem: SagaItem,
+    pattern: Pattern<Action>,
     saga: GeneratorFunction,
     ...args: any[]
 ): ForkEffect;
 export declare function takeLeading(
     sagaItem: SagaItem,
+    pattern: Pattern<Action>,
     saga: GeneratorFunction,
     ...args: any[]
 ): ForkEffect;
 export declare function throttle(
-    ms: number,
     sagaItem: SagaItem,
+    ms: number,
+    pattern: Pattern<Action>,
     saga: GeneratorFunction,
     ...args: any[]
 ): ForkEffect;